backend/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.30
aiomysql>=0.2.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.18
redis[hiredis]>=5.0.0
miniopy-async>=1.21.0
alembic>=1.14.0
httpx>=0.27.0
openpyxl>=3.1.0
pandas>=2.2.0
Pillow>=10.0.0
